Update Topology File To Have Neighbor Specific API Removed
Places neighbor specific items in the neighbors module
Updates unit tests for changes
This allows for the neighbors module to be potentially used in other "applications"
