+++ bcmp/neighbors.c
@@ -1,14 +1,145 @@
 #include "messages/neighbors.h"
+#include "bcmp.h"
 #include "bm_config.h"
 #include "bm_os.h"
+#include "device.h"
+#include "l2.h"
 #include "messages/info.h"
+#include "packet.h"
 #include <inttypes.h>
 #include <string.h>
+
+// Timer to stop waiting for a nodes neighbor table
+#define bcmp_neighbor_timer_timeout_s 1
+#define bcmp_table_max_len 1024
 
 // Pointer to neighbor linked-list
 static BcmpNeighbor *NEIGHBORS;
 static uint8_t NUM_NEIGHBORS = 0;
 NeighborDiscoveryCallback NEIGHBOR_DISCOVERY_CB = NULL;
+NeighborRequestCallback NEIGHBOR_REQUEST_CB = NULL;
+static uint64_t TARGET_NODE_ID = 0;
+BmTimer NEIGHBOR_TIMER = NULL;
+
+/*!
+  @brief Send reply to neighbor table request
+
+  @param *neighbor_table_reply - reply message
+  @param *addr - ip address to send reply to
+  @ret ERR_OK if successful
+*/
+static BmErr bcmp_send_neighbor_table(void *addr) {
+  static uint8_t num_ports = 0;
+  num_ports = bm_l2_get_num_ports();
+  BmErr err = BmENOMEM;
+
+  // Check our neighbors
+  uint8_t num_neighbors = 0;
+  BcmpNeighbor *neighbor = bcmp_get_neighbors(&num_neighbors);
+  uint16_t neighbor_table_len = sizeof(BcmpNeighborTableReply) +
+                                sizeof(BcmpPortInfo) * num_ports +
+                                sizeof(BcmpNeighborInfo) * num_neighbors;
+
+  // TODO - handle more gracefully
+  if (neighbor_table_len > bcmp_table_max_len) {
+    return BmEINVAL;
+  }
+
+  uint8_t *neighbor_table_reply_buff = (uint8_t *)bm_malloc(neighbor_table_len);
+  if (neighbor_table_reply_buff) {
+    memset(neighbor_table_reply_buff, 0, neighbor_table_len);
+
+    BcmpNeighborTableReply *neighbor_table_reply =
+        (BcmpNeighborTableReply *)neighbor_table_reply_buff;
+    neighbor_table_reply->node_id = node_id();
+
+    // set the other vars
+    neighbor_table_reply->port_len = num_ports;
+    neighbor_table_reply->neighbor_len = num_neighbors;
+
+    // assemble the port list here
+    for (uint8_t port = 0; port < num_ports; port++) {
+      neighbor_table_reply->port_list[port].state = bm_l2_get_port_state(port);
+    }
+
+    assemble_neighbor_info_list(
+        (BcmpNeighborInfo *)&neighbor_table_reply->port_list[num_ports],
+        neighbor, num_neighbors);
+
+    err = bcmp_tx(addr, BcmpNeighborTableReplyMessage,
+                  (uint8_t *)neighbor_table_reply, neighbor_table_len, 0, NULL);
+
+    bm_free(neighbor_table_reply_buff);
+  }
+
+  return err;
+}
+
+/*!
+  @brief Handle neighbor table requests
+
+  @param *neighbor_table_req - message to process
+  @param *src - source ip of requester
+  @param *dst - destination ip of request (used for responding to the correct multicast address)
+  @return BmOK if successful
+  @return BmErr if failed
+*/
+static BmErr bcmp_process_neighbor_table_request(BcmpProcessData data) {
+  BmErr err = BmEINVAL;
+  BcmpNeighborTableRequest *request = (BcmpNeighborTableRequest *)data.payload;
+  if (request && ((request->target_node_id == 0) ||
+                  node_id() == request->target_node_id)) {
+    err = bcmp_send_neighbor_table(data.dst);
+  }
+  return err;
+}
+
+/*!
+  @brief Handle neighbor table replies
+
+  @param *neighbor_table_reply - reply message to process
+*/
+static BmErr bcmp_process_neighbor_table_reply(BcmpProcessData data) {
+  BmErr err = BmEINVAL;
+  BcmpNeighborTableReply *reply = (BcmpNeighborTableReply *)data.payload;
+
+  if (TARGET_NODE_ID == reply->node_id) {
+    err = bm_timer_stop(NEIGHBOR_TIMER, 10);
+    if (NEIGHBOR_REQUEST_CB) {
+      bm_err_check(err, NEIGHBOR_REQUEST_CB(reply));
+      NEIGHBOR_REQUEST_CB = NULL;
+    }
+  }
+
+  return err;
+}
+
+/*!
+  @brief Initialize BCMP Topology Module
+
+  @return BmOK on success
+  @return BmErr on failure
+*/
+BmErr bcmp_neighbor_init(void) {
+  BmErr err = BmOK;
+  BcmpPacketCfg neighbor_request = {
+      false,
+      false,
+      bcmp_process_neighbor_table_request,
+  };
+  BcmpPacketCfg neighbor_reply = {
+      false,
+      false,
+      bcmp_process_neighbor_table_reply,
+  };
+
+  bm_err_check(err,
+               packet_add(&neighbor_request, BcmpNeighborTableRequestMessage));
+  bm_err_check(err, packet_add(&neighbor_reply, BcmpNeighborTableReplyMessage));
+
+  return err;
+}
+
 /*!
   @brief Accessor to latest the neighbor linked-list and nieghbor count
 
@@ -273,5 +404,54 @@ void bcmp_neighbor_invoke_discovery_cb(bool discovered,
                                        BcmpNeighbor *neighbor) {
   if (NEIGHBOR_DISCOVERY_CB) {
     NEIGHBOR_DISCOVERY_CB(discovered, neighbor);
+  }
+}
+
+/*!
+  @brief Send neighbor table request to node(s)
+
+  @param target_node_id - target node id to send request to (0 for all nodes)
+  @param *addr - ip address to send to send request to
+  @ret ERR_OK if successful
+*/
+BmErr bcmp_request_neighbor_table(uint64_t target_node_id, const void *addr,
+                                  NeighborRequestCallback request,
+                                  BmTimerCallback timeout) {
+  BmErr err = BmOK;
+  BcmpNeighborTableRequest neighbor_table_req = {.target_node_id =
+                                                     target_node_id};
+  TARGET_NODE_ID = target_node_id;
+  if (NEIGHBOR_TIMER) {
+    bm_timer_delete(NEIGHBOR_TIMER, 10);
+  }
+  NEIGHBOR_TIMER =
+      bm_timer_create("topology_timer", bcmp_neighbor_timer_timeout_s * 1000,
+                      true, NULL, timeout);
+
+  if (NEIGHBOR_TIMER) {
+    err = bm_timer_start(NEIGHBOR_TIMER, 10);
+    bm_err_check(err, bcmp_tx(addr, BcmpNeighborTableRequestMessage,
+                              (uint8_t *)&neighbor_table_req,
+                              sizeof(neighbor_table_req), 0, NULL));
+    NEIGHBOR_REQUEST_CB = request;
+  } else {
+    err = BmENOMEM;
+  }
+
+  return err;
+}
+
+// assembles the neighbor info list
+void assemble_neighbor_info_list(BcmpNeighborInfo *neighbor_info_list,
+                                 BcmpNeighbor *neighbor,
+                                 uint8_t num_neighbors) {
+  uint16_t neighbor_count = 0;
+  while (neighbor != NULL && neighbor_count < num_neighbors) {
+    neighbor_info_list[neighbor_count].node_id = neighbor->node_id;
+    neighbor_info_list[neighbor_count].port = neighbor->port;
+    neighbor_info_list[neighbor_count].online = (uint8_t)neighbor->online;
+    neighbor_count++;
+    // Go to the next one
+    neighbor = neighbor->next;
   }
 }
