+++ common/bm_os.h
@@ -25,6 +25,9 @@ typedef void *BmSemaphore;
 typedef void *BmTimer;
 typedef void *BmTaskHandle;
 
+typedef void (*BmTimerCallback)(BmTimer timer);
+typedef void (*BmTask)(void *arg);
+
 // Memory functions - these may not necessarily be tied to the OS but I'm including them here for now
 void *bm_malloc(size_t size);
 void bm_free(void *ptr);
@@ -42,15 +45,14 @@ BmErr bm_semaphore_give(BmSemaphore semaphore);
 BmErr bm_semaphore_take(BmSemaphore semaphore, uint32_t timeout_ms);
 
 // Task functions
-BmErr bm_task_create(void (*task)(void *), const char *name,
-                     uint32_t stack_size, void *arg, uint32_t priority,
-                     BmTaskHandle task_handle);
+BmErr bm_task_create(BmTask, const char *name, uint32_t stack_size, void *arg,
+                     uint32_t priority, BmTaskHandle task_handle);
 void bm_task_delete(BmTaskHandle task_handle);
 void bm_start_scheduler(void);
 
 // Timer functions
 BmTimer bm_timer_create(const char *name, uint32_t period_ms, bool auto_reload,
-                        void *time_id, void (*callback)(void *));
+                        void *time_id, BmTimerCallback);
 void bm_timer_delete(BmTimer timer, uint32_t timeout_ms);
 BmErr bm_timer_start(BmTimer timer, uint32_t timeout_ms);
 BmErr bm_timer_stop(BmTimer timer, uint32_t timeout_ms);
