+++ test/mocks/mock_neighbors.h
@@ -2,6 +2,7 @@
 #include "messages/neighbors.h"
 #include "util.h"
 
+DECLARE_FAKE_VALUE_FUNC(BmErr, bcmp_neighbor_init);
 DECLARE_FAKE_VALUE_FUNC(BcmpNeighbor *, bcmp_get_neighbors, uint8_t *);
 DECLARE_FAKE_VOID_FUNC(bcmp_check_neighbors);
 DECLARE_FAKE_VOID_FUNC(bcmp_print_neighbor_info, BcmpNeighbor *);
@@ -14,3 +15,7 @@ DECLARE_FAKE_VOID_FUNC(bcmp_neighbor_foreach, NeighborCallback);
 DECLARE_FAKE_VOID_FUNC(bcmp_neighbor_register_discovery_callback,
                        NeighborDiscoveryCallback);
 DECLARE_FAKE_VOID_FUNC(bcmp_neighbor_invoke_discovery_cb, bool, BcmpNeighbor *);
+DECLARE_FAKE_VALUE_FUNC(BmErr, bcmp_request_neighbor_table, uint64_t,
+                        const void *, NeighborRequestCallback, BmTimerCallback);
+DECLARE_FAKE_VOID_FUNC(assemble_neighbor_info_list, BcmpNeighborInfo *,
+                       BcmpNeighbor *, uint8_t);
