+++ test/src/topology_test.cpp
@@ -142,74 +142,6 @@ TEST_F(Topology, topology_start) {
 }
 
 /*!
-  @brief Test obtaining a neighbor table with a request call and processing
-         the reply callback
-*/
-TEST_F(Topology, request_table) {
-  uint32_t addr = (uint32_t)RND.rnd_int(UINT32_MAX, UINT16_MAX);
-  uint64_t node_id = (uint64_t)RND.rnd_int(UINT32_MAX, UINT16_MAX);
-  BcmpNeighborTableReply *reply;
-  NetworkTopology topology;
-  BcmpProcessData data;
-  topology_builder_helper(&topology);
-  reply = (BcmpNeighborTableReply *)bm_malloc(sizeof(BcmpNeighborTableReply) +
-                                              size_info);
-  reply->node_id = node_id;
-  reply->port_len = 2;
-  reply->neighbor_len = 2;
-  memcpy(reply->port_list, replies[0].port_list, size_info);
-  data.payload = (uint8_t *)reply;
-
-  EXPECT_EQ(bcmp_topology_init(), BmOK);
-
-  bm_timer_start_fake.return_val = BmOK;
-  bcmp_tx_fake.return_val = BmOK;
-  EXPECT_EQ(bcmp_request_neighbor_table(node_id, &addr), BmOK);
-
-  // Variable SENT_REQUEST currently un-reachable, cannot test full function
-  packet_process_invoke(BcmpNeighborTableReplyMessage, data);
-
-  // Test failures
-  bcmp_tx_fake.return_val = BmEBADMSG;
-  EXPECT_NE(bcmp_request_neighbor_table(node_id, &addr), BmOK);
-  bm_timer_start_fake.return_val = BmENOMEM;
-  EXPECT_NE(bcmp_request_neighbor_table(node_id, &addr), BmOK);
-  RESET_FAKE(bcmp_tx);
-
-  bm_free(reply);
-  packet_cleanup();
-  topology_builder_cleanup();
-}
-
-/*!
- @brief Test request message callback
-*/
-TEST_F(Topology, request_reply) {
-  BcmpNeighborTableRequest request = {
-      (uint64_t)RND.rnd_int(UINT32_MAX, UINT16_MAX),
-  };
-  BcmpProcessData data;
-  data.payload = (uint8_t *)&request;
-
-  EXPECT_EQ(bcmp_topology_init(), BmOK);
-
-  node_id_fake.return_val = request.target_node_id;
-  bcmp_tx_fake.return_val = BmOK;
-  EXPECT_EQ(packet_process_invoke(BcmpNeighborTableRequestMessage, data), BmOK);
-
-  request.target_node_id = 0;
-  EXPECT_EQ(packet_process_invoke(BcmpNeighborTableRequestMessage, data), BmOK);
-
-  // Test failures
-  bcmp_tx_fake.return_val = BmEBADMSG;
-  EXPECT_NE(packet_process_invoke(BcmpNeighborTableRequestMessage, data), BmOK);
-  node_id_fake.return_val = 1;
-  EXPECT_NE(packet_process_invoke(BcmpNeighborTableRequestMessage, data), BmOK);
-
-  packet_cleanup();
-}
-
-/*!
  @brief Test printing the nodes of a topology
  */
 TEST_F(Topology, print) {
